+++ tests/derive-tests/variant_downcast/fail/enum/tuple_variants/zero_fields.out.rs
@@ -1,6 +1,0 @@
-use enumcapsulate::derive::AsVariant;
-pub struct VariantA;
-pub enum Enum {
-    VariantA(VariantA),
-    VariantB(),
-}