+++ tests/derive-tests/try_into/pass/enum/mixed_variants.out.rs
@@ -9,6 +9,8 @@ pub enum Enum {
     OneStructField { variant: VariantB },
     TwoTupleFields(i32, u32),
     TwoStructFields { a: i32, b: u32 },
+    #[enumcapsulate(exclude)]
+    Excluded(bool),
 }
 impl ::core::convert::TryFrom<Enum> for VariantA {
     type Error = Enum;